refactor(shared-types): transformers の死蔵 API・genre 重複・誤配置を整理
軸2-3。transformers/ の予測不能な命名・死蔵 API・重複を一掃。

- 死蔵 work toFirestore（`as any as WorkDocument` 付き・呼び出しゼロ）を削除し
  workTransformers を fromFirestore のみに
- genre 正規化の4重複（genres/customGenres × 2箇所）を normalizeGenres 1関数に集約し、
  `(g as any).name` の noExplicitAny biome-ignore 4件を解消（unknown narrowing で型安全に）
- audio-button.ts（457→232行）: 変換以外の未使用操作10件（create/update/increment/decrement/
  generateId・本番参照ゼロ）を削除し、conversion（from/toFirestore）だけ残す
- 死蔵 utilities/work-conversions.ts を削除し、Document→Plain の入口を各ドメインの
  *Transformers.fromFirestore に統一（index.ts の JSDoc で正本を明示）。テストの vestigial mock も除去
- work-firestore-final.ts → work-firestore.ts に改名（"final" の由来不明を解消）
- 死蔵一掃で shared-types カバレッジが回復（br 74→78 ほか）したためラチェット引き上げ

shared-types 公開 API 変更（One-Way Door）。削除対象はすべて本番参照ゼロを grep 検証済み。

Closes SPR-197
