+++ apps/web/src/app/creators/[creatorId]/__tests__/page.test.tsx
@@ -58,7 +58,6 @@ vi.mock("@suzumina.click/shared-types", () => ({
 		return types.map((t) => labels[t] || t).join(" / ");
 	}),
 	convertToFrontendWork: vi.fn((work: any) => work), // パススルー変換
-	convertToWorkPlainObject: vi.fn((work: any) => work), // パススルー変換
 }));
 
 // テスト対象のインポート（モック設定後）
