+++ packages/shared-types/src/transformers/audio-button.ts
@@ -225,233 +225,8 @@ function formatDuration(seconds: number): string {
 	return `${minutes}:${secs.toString().padStart(2, "0")}`;
 }
 
-/**
- * Create a new AudioButton with defaults
- */
-export function createAudioButton(params: {
-	buttonText: string;
-	videoId: string;
-	videoTitle: string;
-	startTime: number;
-	endTime: number;
-	tags?: string[];
-	creatorId: string;
-	creatorName: string;
-	isPublic?: boolean;
-}): AudioButton {
-	const duration = params.endTime - params.startTime;
-	const now = new Date().toISOString();
-	const durationText = formatDuration(duration);
-	const searchableText = [
-		params.buttonText.toLowerCase(),
-		params.videoTitle.toLowerCase(),
-		params.creatorName.toLowerCase(),
-		...(params.tags || []).map((tag) => tag.toLowerCase()),
-	].join(" ");
-
-	return {
-		id: "", // Will be set by Firestore
-		buttonText: params.buttonText,
-		videoId: params.videoId,
-		videoTitle: params.videoTitle,
-		startTime: params.startTime,
-		endTime: params.endTime,
-		duration,
-		tags: params.tags || [],
-		creatorId: params.creatorId,
-		creatorName: params.creatorName,
-		isPublic: params.isPublic ?? true,
-		stats: {
-			playCount: 0,
-			likeCount: 0,
-			dislikeCount: 0,
-			favoriteCount: 0,
-			engagementRate: 0,
-		},
-		createdAt: now,
-		updatedAt: now,
-		_computed: {
-			isPopular: false,
-			engagementRate: 0,
-			engagementRatePercentage: 0,
-			popularityScore: 0,
-			searchableText,
-			durationText,
-			relativeTimeText: "たった今",
-		},
-	};
-}
-
-/**
- * Update an AudioButton
- */
-export function updateAudioButton(
-	audioButton: AudioButton,
-	updates: Partial<Pick<AudioButton, "buttonText" | "tags" | "isPublic">>,
-): AudioButton {
-	return {
-		...audioButton,
-		...updates,
-		updatedAt: new Date().toISOString(),
-	};
-}
-
-/**
- * Increment view count
- */
-export function incrementViewCount(audioButton: AudioButton): AudioButton {
-	const newStats = {
-		...audioButton.stats,
-		playCount: audioButton.stats.playCount + 1,
-		engagementRate: calculateEngagementRate(
-			audioButton.stats.playCount + 1,
-			audioButton.stats.likeCount,
-			audioButton.stats.dislikeCount,
-		),
-	};
-
-	return {
-		...audioButton,
-		stats: newStats,
-		updatedAt: new Date().toISOString(),
-	};
-}
-
-/**
- * Increment like count
- */
-export function incrementLikeCount(audioButton: AudioButton): AudioButton {
-	const newStats = {
-		...audioButton.stats,
-		likeCount: audioButton.stats.likeCount + 1,
-		engagementRate: calculateEngagementRate(
-			audioButton.stats.playCount,
-			audioButton.stats.likeCount + 1,
-			audioButton.stats.dislikeCount,
-		),
-	};
-
-	return {
-		...audioButton,
-		stats: newStats,
-		updatedAt: new Date().toISOString(),
-	};
-}
-
-/**
- * Decrement like count
- */
-export function decrementLikeCount(audioButton: AudioButton): AudioButton {
-	const newStats = {
-		...audioButton.stats,
-		likeCount: Math.max(0, audioButton.stats.likeCount - 1),
-		engagementRate: calculateEngagementRate(
-			audioButton.stats.playCount,
-			Math.max(0, audioButton.stats.likeCount - 1),
-			audioButton.stats.dislikeCount,
-		),
-	};
-
-	return {
-		...audioButton,
-		stats: newStats,
-		updatedAt: new Date().toISOString(),
-	};
-}
-
-/**
- * Increment dislike count
- */
-export function incrementDislikeCount(audioButton: AudioButton): AudioButton {
-	const newStats = {
-		...audioButton.stats,
-		dislikeCount: audioButton.stats.dislikeCount + 1,
-		engagementRate: calculateEngagementRate(
-			audioButton.stats.playCount,
-			audioButton.stats.likeCount,
-			audioButton.stats.dislikeCount + 1,
-		),
-	};
-
-	return {
-		...audioButton,
-		stats: newStats,
-		updatedAt: new Date().toISOString(),
-	};
-}
-
-/**
- * Decrement dislike count
- */
-export function decrementDislikeCount(audioButton: AudioButton): AudioButton {
-	const newStats = {
-		...audioButton.stats,
-		dislikeCount: Math.max(0, audioButton.stats.dislikeCount - 1),
-		engagementRate: calculateEngagementRate(
-			audioButton.stats.playCount,
-			audioButton.stats.likeCount,
-			Math.max(0, audioButton.stats.dislikeCount - 1),
-		),
-	};
-
-	return {
-		...audioButton,
-		stats: newStats,
-		updatedAt: new Date().toISOString(),
-	};
-}
-
-/**
- * Increment favorite count
- */
-export function incrementFavoriteCount(audioButton: AudioButton): AudioButton {
-	const newStats = {
-		...audioButton.stats,
-		favoriteCount: audioButton.stats.favoriteCount + 1,
-	};
-
-	return {
-		...audioButton,
-		stats: newStats,
-		updatedAt: new Date().toISOString(),
-	};
-}
-
-/**
- * Decrement favorite count
- */
-export function decrementFavoriteCount(audioButton: AudioButton): AudioButton {
-	const newStats = {
-		...audioButton.stats,
-		favoriteCount: Math.max(0, audioButton.stats.favoriteCount - 1),
-	};
-
-	return {
-		...audioButton,
-		stats: newStats,
-		updatedAt: new Date().toISOString(),
-	};
-}
-
-/**
- * Generate a unique ID (for client-side use only)
- */
-export function generateId(): string {
-	return Math.random().toString(36).substring(2) + Date.now().toString(36);
-}
-
 // Re-export as namespace for backward compatibility
 export const audioButtonTransformers = {
 	fromFirestore,
 	toFirestore,
-	createAudioButton,
-	updateAudioButton,
-	incrementViewCount,
-	incrementLikeCount,
-	decrementLikeCount,
-	incrementDislikeCount,
-	decrementDislikeCount,
-	incrementFavoriteCount,
-	decrementFavoriteCount,
-	generateId,
 };
